pyproject.toml
[tool.poetry]
name = "taskfy"
version = "0.1.0"
description = ""
authors = ["Thiago Regueira <thiago.regueira@yahoo.com.br>"]
readme = "README.md"
packages = [{ include = "taskfy" }]

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.1.11"
httpx = "^0.26.0"
isort = "^5.13.2"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['migrations']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn taskfy.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=taskfy -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
